+++ test/crossbeams/link_test.rb
@@ -3,7 +3,7 @@ require 'test_helper'
 class Crossbeams::LinkTest < Minitest::Test
   def test_defaults
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/')
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
   end
 
   def test_invalid_args
@@ -17,16 +17,16 @@ class Crossbeams::LinkTest < Minitest::Test
 
   def test_styles
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', style: :link)
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', style: :button)
-    assert_equal '<a href="/" class="f6 link dim br2 ph3 pv2 dib white bg-silver">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" class="f6 link dim br2 ph3 pv2 dib white bg-silver" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', style: :small_button)
-    assert_equal '<a href="/" class="link dim br1 ph2 dib white bg-silver">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" class="link dim br1 ph2 dib white bg-silver" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', style: :action_button)
-    assert_equal '<a href="/" class="f6 link dim br2 ph3 pv2 dib white bg-green">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" class="f6 link dim br2 ph3 pv2 dib white bg-green" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', style: :back_button)
     assert_match(/<svg class=["|']cbl-icon["|']/, renderer.render)
@@ -42,15 +42,15 @@ class Crossbeams::LinkTest < Minitest::Test
 
   def test_grid
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/')
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', grid_id: 'a_grid_id')
-    assert_equal '<a href="/" data-grid-id="a_grid_id">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-grid-id="a_grid_id" data-new-page-link="true">ClickMe</a>', renderer.render.strip
   end
 
   def test_prompt
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/')
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', prompt: 'Is this OK?')
     assert_equal '<a href="/" data-prompt="Is this OK?">ClickMe</a>', renderer.render.strip
@@ -59,7 +59,7 @@ class Crossbeams::LinkTest < Minitest::Test
     assert_equal '<a href="/" data-prompt="Are you sure?">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', prompt: false)
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', prompt: 'y')
     assert_equal '<a href="/" data-prompt="Are you sure?">ClickMe</a>', renderer.render.strip
@@ -76,17 +76,17 @@ class Crossbeams::LinkTest < Minitest::Test
 
   def test_visible
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', visible: false)
-    assert_equal '<a href="/" hidden>ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" hidden data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', visible: true)
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
 
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', visible: 'ANY NON-FALSE VALUE')
-    assert_equal '<a href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
   end
 
   def test_id
     renderer = Crossbeams::Layout::Link.new(text: 'ClickMe', url: '/', id: 'an_id')
-    assert_equal '<a id="an_id" href="/">ClickMe</a>', renderer.render.strip
+    assert_equal '<a id="an_id" href="/" data-new-page-link="true">ClickMe</a>', renderer.render.strip
   end
 end