+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Added
 - A `CallbackSection` can be rendered to work inside a dialog by calling the `remote!` method on the section.
 ### Changed
+- Set `data-new-page-link` on `Link`s when the link opens a new page (so that it can be disabled in the UI once clicked)
 ### Fixed
 
 ## [0.6.0] - 2022-08-28